[ADD] dolmen: Cash Menu and Payment from a SO
[ADD] dolmen: Cash Menu

Adding a new menu, so the cashiers can access more quickly to the bank
statements with the type `cash`.

[ADD] dolmen: General Public (Contact)

Adding a new field `is_general_public` in the `res.partner` model
to allow us know if a contact doesn't have a VAT or if the VAT
is a generic one.

[ADD] dolmen: Payment from Sale Order

Adding a new button "Register Payment" in the sale order form model,
that will allow the user to generate a "payment" (transaction `bank.statement.line`)
from a sale order that is not invoiced and then the user will have to
do the reconciliation of the payments when the invoice is made.